backend/requirements.txt
fastapi
uvicorn
python-dotenv
torch
torchvision
pillow
